Extracted all net.Conn code to separate go file.
Created peerConn struct which should do all tcp reads
and writes. Before any read and write it sets timeouts on
underlaying tcp connection.

Will see if there is point in adding cancelation to it using ctx.
If no, then remove ctx from method signatures as we are not
using them.